[PLAT-18674] Add extra warnings to continuous backup restore flow
Summary:
This diff adds extra warnings to continuous backup restore flow.
When restoring platform backup on a platform with one or more universes, we now show a confirmation
modal to the user which warns about data loss from inconsistent state.
Also in the restore form modal, we now show a warning banner at the bottom which
calls out the platform shutdown and restart sequence.

Test Plan:
Attempt platform restore on a platform with universes and on a platform without universes.
Verify the warning banner UI is as expected.
{F401858}
{F403817}
{F403818}

Reviewers: rmadhavan

Reviewed By: rmadhavan

Differential Revision: https://phorge.dev.yugabyte.com/D47304