[#26734] YSQL: Validate tablet schema in tests after RollbackToSubTransaction operation
Summary:
As part of D45506 / 30a37d3, we introduced logic to store the sub_transaction_id as part of every DDL operation.

This is a follow-up from that revision to handle an unresolved code review comment:
```
What is the test plan for seeing if we got the tablet schema changes correct? I didn't see anything in the tests that seem to test this.
```

This revision adds validation on the tablet schema after each rollback to savepoint operation. We also validate the `previous_schema` field to ensure that the correct schema gets tracked in the verifier state. The `previous_schema` field is used during rollback to issue an `ALTER TABLE` operation.
Jira: DB-16109

Test Plan:
Jenkins: test regex: .*PgDdlSavepointMiniClusterTest.*

./yb_build.sh --cxx-test pgwrapper_pg_ddl_transaction-test

Reviewers: mlillibridge

Reviewed By: mlillibridge

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D47147