[#28921] docdb: Add metrics for initial transaction locality
Summary:
We currently expose a transaction_promotions metric to track transaction promotions. But this does not give visibility into whether a workload is using global transactions or not, since we also try to start transactions as global whenever we know it's needed, and such transactions do not involve a transaction promotion.

Additionally, we don't expose transaction locality anywhere explicitly except via vlogs, so combined with the above, it is difficult to verify that global transactions are or are not being used except indirectly via commit latency.

This change adds the following server-level metrics:
- initially_global_transactions - number of transactions started as global (including transactions in transaction pool)
- initially_region_local_transactions - number of transactions started as region-local (including transactions in transaction pool)
- initially_tablespace_local_transactions - number of transactions started as tablespace-local (including transactions in transaction pool)

Additionally, a bug where the transaction_promotions counter was reset after inactivity due to its lifespan being tied to presence of live YBTransaction objects was fixed by tying the counter lifespan to TransactionManager instead.
Jira: DB-18646

Test Plan: Jenkins. Also verified metrics on local cluster.

Reviewers: bkolagani

Reviewed By: bkolagani

Subscribers: schandra, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D47401