[#28000] YSQL: Fix PgHintTableTest.PreparedStatementHintCacheRefresh with txn ddl
Summary:
This revision fixes the test PgHintTableTest.PreparedStatementHintCacheRefresh with transactional DDL.

When transactional DDL is enabled, an insert into the `hint_plan.hints` table leads to a single cache miss which is different from the case when txn
ddl is disabled. This change in behavior is because the invalidation messages are received only on commit and when txn ddl is enabled, all 3 queries
(insert and 2 as part of yb_increment_db_catalog_version_with_inval_messages) are done in the same transaction.
Jira: DB-17619

Test Plan:
Jenkins: test regex: .*PgHintTableTest.*

Enabled transactional DDL and ran

./yb_build.sh --cxx-test pgwrapper_pg_hint_table-test --gtest_filter PgHintTableTest.PreparedStatementHintCacheRefresh

Reviewers: kfranz, myang

Reviewed By: kfranz

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D47416