[PLAT-18735][PLAT-18736] Ensure scrape interval cannot be greater than scrape timeout
Summary:
This diff fixes 2 tickets:

1. [PLAT-18735] Ensure scrape interval cannot be greater than scrape timeout
2. [PLAT-18736] Remove sleep between servers timeout from DBME API

This adds more validation and removes unnecessary fields from the API which are not used and required.
The field `SleepAfterRestartSchema` contains `sleep_after_master_restart_millis` and `sleep_after_tserver_restart_millis`, both of which are not applicable to this API. This is a new API that was introduced in 2025.2, which has not yet been released, so we can remove it without deprecation as this fix will be backported to 2025.2.

Test Plan:
Manually tested via the API. Got errors for the following cases:
1. `scrape_interval_seconds=10, scrape_timeout_seconds=20`
2. `scrape_interval_seconds=-30, scrape_timeout_seconds=20`
3. `scrape_interval_seconds=30, scrape_timeout_seconds=-20`

Run UTs.
Run itests.

Reviewers: amalyshev, dkumar

Reviewed By: amalyshev

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D47419