[#28868] ASH: Fix flaky test PgAshTest.TestTServerMetadataSerializer
Summary:
This diff fixes flaky test PgAshTest.TestTServerMetadataSerializer
in release builds.

The test was likely flaky because the inserts happen too quickly in
release builds and there was not enough time to sample. Changing
the test to use a long insert will fix this. The number of values to
insert will depend on the build type i.e. 100000 rows on debug builds
and 10000000 rows on release builds.
Jira: DB-18586

Test Plan:
Jenkins: test regex: .*TestTServerMetadataSerializer.*

./yb_build.sh release --cxx-test pg_ash-test --gtest_filter PgAshTest.TestTServerMetadataSerializer -n 10

Reviewers: ishan.chhangani

Reviewed By: ishan.chhangani

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D47418