[#28936] docdb: Disable object locking for system.transaction creation tests
Summary:
GlobalTransactionTableCreationTest.CreateGlobalTransactionTableAfterFailover and
CreateTableITest.TestTransactionStatusTableCreation fail with table locks enabled.

Both tests involve a setup where we start master/tserver and wait for them to be ready,
before system.transactions has been created. For YSQL, this happens at startup normally, but
for YCQL, this is delayed until a table with transactions enabled is created.

The two tests are specifically testing behavior around the creation of system.transactions at a
slightly delayed time, by initializing the cluster in a configuration where it is not initially creatable
(insufficient tservers, tservers with bad placement). The tests verify that transactions are not usable,
then move to cluster into a better state where system.transactions should be created, and test
that transactions are now usable.

But when table locks are enabled, we get stuck waiting for the cluster to even become ready to
accept connections.

Disabling table locks for this test until the underlying issue is fixed.
Jira: DB-18664

Test Plan: Jenkins: test regex: GlobalTransactionTableCreationTest.CreateGlobalTransactionTableAfterFailover|CreateTableITest.TestTransactionStatusTableCreation

Reviewers: bkolagani, rthallam

Reviewed By: rthallam

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D47471