[BACKPORT 2025.1][#27288] YSQL: Avoid 'restart read required' error during yb_index_check()
Summary:
yb_index_check() scans all the rows of the index and the base relation. Hence, it becomes prone to 'Restart read required' error if any row is involved in a write operation concurrently with the yb_index_check() read point, which is a high probability event.

This hurts usability, but more importantly, yb_index_check() is not concerned with read-after-commit-visibility. It picks up a read time (and an associated snapshot) and uses it to scan both the index and base relation. Even if a write that got committed before the scan read point (as observed by an external clock) is missing from the snapshot due to clock skew, that's acceptable—its effects will be absent from both the index and the base table scans.

So, it is safe to set GUC yb_read_after_commit_visibility to 'relaxed' while executing yb_index_check() to avoid these errors. Note that this GUC cannot be set from within a transaction block. Don't do it in such cases and throw a warning instead.
Jira: DB-16780

Original commit: 841a5634ae33ce62e28d858df889b151cbefc783 / D46002

Test Plan: ./yb_build.sh --gtest_filter PgYbIndexCheckTest.YbIndexCheckRestartReadRequired

Reviewers: patnaik.balivada

Reviewed By: patnaik.balivada

Subscribers: yql, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D47513