[#27948] YSQL: Allow DELETE and (re-) INSERT of a row to be executed in a single flush
Summary:
### The problem
Users can define expression indexes on subkeys of JSON columns. More specifically, the `->>` operator (which is used to access subkeys of JSON columns) casts subkeys to text, which is supported for indexing.
Updates to such indexes are expensive. Consider the following setup:
```
CREATE TABLE t_json (k INT PRIMARY KEY, v INT, object JSONB);

-- create an expression index on object["b"]
CREATE UNIQUE INDEX idx_json ON t_json ((object->>'b'));

-- update (set) object["a"] to 100
UPDATE t_json SET object = jsonb_set(object, '{a}', to_jsonb(100)) WHERE k = 1;

-- EXPLAIN plan
                                                QUERY PLAN
                              QUERY PLAN
----------------------------------------------------------------------
 Update on t_json (actual rows=1 loops=1)
   Storage Table Write Requests: 1
   Storage Index Write Requests: 2
   ->  Index Scan using t_json_pkey on t_json (actual rows=1 loops=1)
         Index Cond: (k = 1)
         Storage Table Read Requests: 1
         Storage Table Rows Scanned: 1
 Storage Read Requests: 1
 Storage Rows Scanned: 1
 Storage Write Requests: 3                   <---- NOTICE THIS
 Storage Flush Requests: 2                   <---- AND THIS
(11 rows)
```

The execution of the above query involves two index write requests  to `idx_json` which do the following:
 - DELETE row corresponding to `object->>’b’ = 20`
 - INSERT row corresponding to `object->>’b’ = 20`

Updates to the index row are scheduled even though the index row remains unchanged and is logically unaffected by the update.
This is because the planner is unaware of the underlying structure of the column `object`.
Since both write operations reference the same row, they are required to be executed in separate flushes, making the update expensive.

### The solution
In order to execute both DELETE and INSERT operations in a single flush, the following requirements must hold:
 - The order in which the operations are applied to memtable must be deterministic  - This is guaranteed by the tserver today.
 - If the delete fails, the insert must be aborted - This is guaranteed by transaction atomicity.
 - The uniqueness check for the INSERT must be skipped - This is not guaranteed today.

Non-unique indexes do NOT require the index row that is being inserted to be unique. Therefore, the insert operation for these indexes use UPSERT semantics.
Unique indexes require the index row to be unique. However, in the context of this problem, the insert operation is always preceded by a delete.
Therefore it is guaranteed that insert the index row will never return a unique constraint violation.
This implies that the insert operation for unique indexes can also use UPSERT semantics when accompanied by a delete operation in the same flush.

This revision:
 - Relaxes the rule to disallow flushing two write operations to the same row (key) if the two operations are a DELETE and an UPSERT (in that order)
 - Adds logic to convert a DocDB INSERT operation into a DocDB UPSERT operation for unique indexes when a DocDB DELETE operation is already enqueued.

### Flags
This revision introduces a new pggate gflag `ysql_optimize_index_row_rewrites` which when set to true enables the above optimization.
The flag is enabled (true) by default.
This flag does NOT require a corresponding auto-flag or any upgrade protection.

### Sample Output
This revision produces the following output for the above query:
```
                              QUERY PLAN
----------------------------------------------------------------------
 Update on t_json (actual rows=1 loops=1)
   Storage Table Write Requests: 1
   Storage Index Write Requests: 2
   ->  Index Scan using t_json_pkey on t_json (actual rows=1 loops=1)
         Index Cond: (k = 1)
         Storage Table Read Requests: 1
         Storage Table Rows Scanned: 1
 Storage Read Requests: 1
 Storage Rows Scanned: 1
 Storage Write Requests: 3                   <---- NOTICE THIS
 Storage Flush Requests: 1                   <---- AND THIS
(11 rows)
```
Jira: DB-17563

Test Plan:
Run the following test:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressUpdateOptimized#schedule'
```

Reviewers: sergei, hsunder, rthallam, dmitry

Reviewed By: dmitry

Subscribers: dmitry, yql, smishra

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D45427