[PLAT-18751]DB version changes unexpectedly when configuring read replicas in YBA and coming back to Primary Cluster page.
Summary:
Previously, the software version field was being reset when switching from the async cluster to the primary cluster. This happened because the field was being reset whenever osPatching was enabled and the field is disabled, without checking if the DB version was already set. We’ve now added that check to prevent unintended resets.

I’ve also verified that the software version field is still correctly reset when the CPU architecture is changed.

Test Plan: Tested manually

Reviewers: lsangappa

Reviewed By: lsangappa

Differential Revision: https://phorge.dev.yugabyte.com/D47540