[#28438] docdb: Fix CQL server shutdown deadlock
Summary:
The `CQLServiceImpl` class has a background task inside `SystemQueryCache`. This task does some IO to fetch YSQL data. There is a race with the shutdown thread in which this background task will get stuck waiting for the CQL processor to invoke the callback indefinitely, blocking the shutdown flow.

This diff fixes the deadlock by having the background task wait on two signals. One is the callback passed to the CQL processor, the other is shutdown.

This diff also enables graceful shutdown in all CQL java tests. Graceful shutdown had been disabled because of how often it blocks when CQL is enabled.

Test Plan:
```
./yb_build.sh fastdebug --java-test 'org.yb.cql.TestBlobDataType' -n 100 --stop-at-failure
```

Reviewers: bkolagani

Reviewed By: bkolagani

Subscribers: yql, ybase, rthallam

Differential Revision: https://phorge.dev.yugabyte.com/D47195