[#28978] YSQL: Increase timeout for waiting for catalog version
Summary:
In one of the v8 tests with 2024.2.5.1, the following log was observed:

```
2025-10-16 10:29:40.878 UTC [78993] ERROR:  Failed to find suitable shared memory index for db 16643: Timed out (yb/util/backoff_waiter.cc:78): Operation 'Database 16643 is not ready in Yugabyte shared memory' didn't complete within 10000ms, there may be too many databases or the database might have been dropped: Operation 'Database 16643 is not ready in Yugabyte shared memory' didn't complete within 10000ms
```

This indicates that a newly created database 16643 was created. Since it was created, a new row is
inserted into the table pg_yb_catalog_version table, but each tserver needs to receive the next
heartbeat response from the master to pass back the new catalog version info that includes the new
database 16643. In the v8 test environment, it has limited resources so it appears that both the
master leader and 2 of the tservers are overloaded. As a result, master did not send back the
expected heartbeat response that included the catalog version for database 16643 until 10 seconds
later. By that time the connection wait time has already timed out.

This has come up intermittently, especially in some stress test scenarios. I increased the time out
to 30 seconds, which should practically eliminate all such time out. If it still happened, it means
that the cluster is so overloaded, and perhaps it is good to reject a new connection anyway.

I also increased another similar timeout from 20s to 30s.
Jira: DB-18722

Test Plan: ./yb_build.sh release --cxx-test pg_catalog_version-test --gtest_filter PgCatalogVersionTest.SimulateDelayedHeartbeatResponse

Reviewers: kfranz, sanketh

Reviewed By: sanketh

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D47555