[BACKPORT 2025.1][#28828] docdb: Perform initial is-region-local checks on tserver side
Summary:
When using region local transactions, there are two times we check for region locality:
1. We perform a check in postgres code (YBCIsRegionLocal) that takes the relation's tablespace oid,
   calculates the tablespace cost, and checks this is at most the interzone cost. This value is
   propagated up through pggate and eventually passed to the tserver in the Perform RPC, where it is
   checked only if the Perform RPC starts a transaction.
2. We perform checks at the tablet level in YBTransaction to decide whether we need to promote the
   the transaction to global.

(1) does nothing when table-level locking is enabled, because the transaction is now always started
before the Perform RPC (always done during an AcquireObjectLocks RPC), forcing us to use (2)
instead, which has performance implications (#28317).

Additionally, as part of work for tablespace locality, we now pass tablespace oid for read/write
operations inside the Perform, and additionally have tablespace information on the tserver for any
tablespace with transaction tables. So the calculation for (1) can be done entirely on tserver side.

This revision adds calculation of region locality on the tserver side using the above information.
A future change will address #28317 by pushing tablespace oid up in AcquireObjectLocks RPC and
reusing this logic.

The old logic (YBCIsRegionLocal) is left intact for now to avoid severe performance regressions
during upgrade, since the new logic only works after finalizing an upgrade to a version with the
enable_tablespace_based_transaction_placement autoflag (this controls availability of tablespace
information on the tserver). It can be removed once we have a guarantee that all upgrades come
from a verison with this autoflag (2025.1.1.1/2025.1.2/2025.2.0+).

There is also a potential for performance regression if the auto_create_transaction_tables is off
(since the new calculation is based off of tablespaces with transaction tables), but this flag is on
by default and turning it off is ill-advised in general (manual creation of transaction tables via
yb-admin becomes required).
Jira: DB-18538

Original commit: 321857cbfc097f326da8fe20f254d4af1d6124d2 / D47241

Test Plan: Jenkins

Reviewers: bkolagani

Reviewed By: bkolagani

Subscribers: yql, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D47574