[#28837] cdc: replace TEST_ApplyAndFlush call with shutdown aware waiter
Summary:
The `TEST_ApplyAndFlush` call in `cdc_state_table` is on the call path of the background `UpdatePeersAndMetrics` thread in the `CDCStateService`. This call can block graceful shutdown if the background thread hits it while shutting down. This diff replaces the `TEST_ApplyAndFlush` call with logic to wait either the callback is called or the cdc service is shut down.
Jira: DB-18549

Test Plan:
```
./yb_build.sh release --java-test 'org.yb.cdc.ysql.TestArrayTypes#testArrayTypes' -n 100 --stop-at-failure
```

Reviewers: bkolagani, stiwary

Reviewed By: bkolagani

Subscribers: ybase, rthallam

Differential Revision: https://phorge.dev.yugabyte.com/D47523