[docs] Change <IP#> to 127.0.0.# (#28853)
Some commands use <IP#>.  That requires the user to substitute it with
something meaningful.  Rather, use 127.0.0.# so that it works for the
local machine, and it is still obvious to the user to substitute it with
other IPs for multi-machine setups.

Apply not just to preview but also all prior versions.  This involves
- some trivial merge conflicts due to whitespace differences by commits
  a565e7f9cdaa06b21da893e5e01c5a5b094fe46f and
  8db3d6fe991a3420364b431a763db76a60e3c682
- introducing the whole yugabyted commands section in v2.20's
  tablespaces.md since it was somehow missing