[#28972] docdb: Skip graceful shutdown in pg regress tests
Summary:
The test TestPgRegressLargeTable is currently broken. There is a separate root cause affecting this test. But for now disable graceful shutdown during this test so it is easier to debug the current root cause of this test's failures.

This diff also introduces a helper function to capture calculations made for different built max runtimes.

Test Plan: n/a

Reviewers: patnaik.balivada

Reviewed By: patnaik.balivada

Subscribers: ybase, rthallam

Differential Revision: https://phorge.dev.yugabyte.com/D47546